refactor: reorganize code + delete duplicate files
