+++ model/dependencies.txt
@@ -1,4 +1,4 @@
 MATLAB	9.12.0.2039608 (R2022a) Update 5
 libSBML	5.19.0
-RAVEN_toolbox	commit a09ece3
+RAVEN_toolbox	2.8.4
 COBRA_toolbox	commit 28acb94
